FastAPI学习/WebAIPilot/requirements.txt
colorlog==6.10.1
playwright==1.56.0
PyYAML==6.0.3
